fix(tts): report canonical measured audiobook save duration
The save response reported audio_duration_sec from a per-chunk f32 running
accumulation, which drifts from the canonical f64 total the manifest
persists via build_playback_index over long books.

write_manifest now returns its measured (audio_duration_sec, wav_bytes)
totals, and the final saved event plus the save response use those
WAV-header-derived values instead of the running f32 sum. The response
field widened from f32 to f64 (transparent over IPC; TypeScript reads it
as a number either way). Live mid-save progress stays f32 since it is
display-only.

Callers are unaffected: the bundle import path discards the returned
tuple and the stub path is unchanged.

+++ src-tauri/src/native_tts/types.rs
@@ -306,7 +306,10 @@ pub(crate) struct NativeAudiobookSaveResponse {
     pub(crate) complete: bool,
     pub(crate) dir: String,
     pub(crate) generate_ms: u128,
-    pub(crate) audio_duration_sec: f32,
+    // Canonical total measured from WAV headers in `build_playback_index`, so the
+    // reported duration matches the persisted manifest instead of drifting from a
+    // per-chunk f32 accumulation over a long book.
+    pub(crate) audio_duration_sec: f64,
     pub(crate) wav_bytes: usize,
     pub(crate) applied_thread_count: i32,
     pub(crate) backend: String,
